This commit was manufactured by cvs2svn to create tag 'committed'.
git-svn-id: svn://svn.zabbix.com/tags/committed@3125 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/active.h
@@ -32,7 +32,9 @@ METRIC
 	int	refresh;
 	int	nextcheck;
 	int	status;
-	int	lastlogsize;
+/* Must be long for fseek() */
+/*	int	lastlogsize;*/
+	long	lastlogsize;
 };
 
 pid_t   child_active_make(int i,char *server, int port);
