+++ src/zabbix_agent_win32/xml.cpp
@@ -36,7 +36,7 @@ INIT_CHECK_MEMORY(main);
 
 	if(ret == SUCCEED)
 	{
-		len=end-(start+strlen(tag_open));
+		len = (int)(end - (start + (int)strlen(tag_open)));
 
 		if(len>maxlen)	len=maxlen;
